Move all workflow stuff into the Workflows subdirectory.
This is similar to how it's done in OrchardCore, e.g. in the OrchardCore.Contents module.
