+++ Lombiq.HelpfulExtensions/Extensions/ContentSets/Startup.cs
@@ -1,6 +1,4 @@
-﻿using Lombiq.HelpfulExtensions.Extensions.Activities;
-using Lombiq.HelpfulExtensions.Extensions.ContentSets.Drivers;
-using Lombiq.HelpfulExtensions.Extensions.ContentSets.Events;
+﻿using Lombiq.HelpfulExtensions.Extensions.ContentSets.Drivers;
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.Indexes;
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.Models;
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.Services;
@@ -10,7 +8,6 @@ using Microsoft.Extensions.DependencyInjection;
 using OrchardCore.ContentManagement;
 using OrchardCore.ContentManagement.Display.ContentDisplay;
 using OrchardCore.Modules;
-using OrchardCore.Workflows.Helpers;
 using System;
 
 namespace Lombiq.HelpfulExtensions.Extensions.ContentSets;
@@ -32,16 +29,5 @@ public class Startup : StartupBase
     public override void Configure(IApplicationBuilder app, IEndpointRouteBuilder routes, IServiceProvider serviceProvider)
     {
         // No need for anything here yet.
-    }
-}
-
-[RequireFeatures("OrchardCore.Workflows")]
-public class WorkflowStartup : StartupBase
-{
-    public override void ConfigureServices(IServiceCollection services)
-    {
-        services.AddActivity<ContentSetGetSupportedOptionsEvent, ContentSetGetSupportedOptionsEventDisplayDriver>();
-        services.AddActivity<ContentSetCreatingEvent, ContentSetCreatingEventDisplayDriver>();
-        services.AddScoped<IContentSetEventHandler, WorkflowContentSetEventHandler>();
     }
 }
