+++ Lombiq.HelpfulExtensions/Extensions/ContentSets/Workflows/Handlers/WorkflowContentSetEventHandler.cs
@@ -1,6 +1,7 @@
-﻿using Lombiq.HelpfulExtensions.Extensions.Activities;
+﻿using Lombiq.HelpfulExtensions.Extensions.ContentSets.Events;
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.Models;
 using Lombiq.HelpfulExtensions.Extensions.ContentSets.ViewModels;
+using Lombiq.HelpfulExtensions.Extensions.ContentSets.Workflows.Activities;
 using Lombiq.HelpfulLibraries.OrchardCore.Workflow;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
@@ -11,7 +12,7 @@ using OrchardCore.Workflows.Services;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Lombiq.HelpfulExtensions.Extensions.ContentSets.Events;
+namespace Lombiq.HelpfulExtensions.Extensions.ContentSets.Workflows.Handlers;
 
 public class WorkflowContentSetEventHandler : IContentSetEventHandler
 {
